requirements.txt
crewai==0.28.0
langchain==0.0.335
python-dotenv==1.0.0
pytest==7.4.3
pytest-mock==3.12.0
pyyaml==6.0.1
beautifulsoup4==4.12.2
requests==2.31.0 